requirements.txt
Django==2.2.3
django-crispy-forms==1.7.2
django-recaptcha==2.0.4
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
